bridge: refactor with new profile module
