+++ ios/Classes/Interop/Action.swift
@@ -17,9 +17,10 @@ enum Action: String {
     case user_getInstallationID = "user.getInstallationID"
     case user_fetchAttributes = "user.fetch.attributes"
     case user_fetchTags = "user.fetch.tags"
-    case user_edit = "user.edit"
-    case user_trackEvent = "user.track.event"
-    case user_trackLocation = "user.track.location"
+    
+    case profile_edit = "profile.edit"
+    case profile_trackEvent = "profile.track.event"
+    case profile_trackLocation = "user.track.location"
     
     case inbox_createInstallationFetcher = "inbox.createInstallationFetcher"
     case inbox_createUserFetcher = "inbox.createUserFetcher"
